kiss: Always remove charset.alias
We never want this. KISS_RM used to remove it
unconditionally but seeing as this is no longer
the case, let's do it post-build.

This is a file related to intltool/gettext and
is installed by multiple packages under the same
name causing endless conflicts.

We have no use for this and nor do I want to
handle this kind of file installation in the
package manager.

It's ugly and I'm glad it's gone.

FossilOrigin-Name: d74ce8be8667e38eb9b710efd2f65e27e085110adbea2346cf6fee1dfff85778