kiss-repodepends: don't output the dependencies of installed package
kiss-repodepends did not honour the user's KISS_PATH and outputted
the last path it found. If the user has the package installed, the
last path will always be the system database, which defeats the whole
purpose of "repodepends". This patch will use the first path it finds,
which will be the user's preferred repository in the KISS_PATH.