kiss: switch to an 'extension' argument for listing extensions.
Extensions used too much screen space. If a user has their own
extensions it only gets bigger. kiss extensions are not as important
as the package manager itself, so the user shouldn't be required to
scroll up everytime they want to list options. If a user wants to
be informed about the extensions they can now pass an 'e' or
'extension' argument to kiss.