kiss-stat: fix segfault
This fixes an issue when the file doesn't have a registered owner in
the passwd database, but we still tried to print it. We now check if
the entry exists before we try printing it.

FossilOrigin-Name: f34d077df5b31e7fc810e6b32e8fd526962e1268544661c535477eb8ff40db60